Parallelize NDPluginStats with OpenMP
This includes a parallelization in the image level processed by the
plugins, which is different from the current task queue used as a
producer and consumer system.

OpenMP has been used since it is quite easy to define the
parallelization while being portable to many systems. These changes have
not been tested in operating systems other than GNU/Linux.

NDPluginProfile has not been parallelized, since it did not show to be
the bottleneck in performance in a profiling made with help of Linux
Perf tool with instruction count measurement.

Co-authored-by: Henrique F. Simoes <henrique.simoes@students.ic.unicamp.br>